requirements.txt
langdetect
selenium
datetime
asyncio
schedule
telethon
requests
tqdm
pyTelegramBotAPI
python-dotenv